(FACT-1145) Add virtualization resolver for OpenBSD
This also moves get_product_name_vm() into the base resolver
