+++ lib/src/facts/openbsd/collection.cc
@@ -5,6 +5,7 @@
 #include <internal/facts/openbsd/dmi_resolver.hpp>
 #include <internal/facts/openbsd/memory_resolver.hpp>
 #include <internal/facts/openbsd/networking_resolver.hpp>
+#include <internal/facts/openbsd/virtualization_resolver.hpp>
 #include <internal/facts/posix/identity_resolver.hpp>
 #include <internal/facts/posix/kernel_resolver.hpp>
 #include <internal/facts/posix/ssh_resolver.hpp>
@@ -28,6 +29,7 @@ namespace facter { namespace facts {
         add(make_shared<openbsd::networking_resolver>());
         add(make_shared<openbsd::dmi_resolver>());
         add(make_shared<openbsd::memory_resolver>());
+        add(make_shared<openbsd::virtualization_resolver>());
     }
 
 }}  // namespace facter::facts
